Isolate renderer API from global GL headers
