+++ Quake/gl_fog.c
@@ -22,6 +22,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 //gl_fog.c -- global fog
 
 #include "quakedef.h"
+#include "glquake.h"
 
 //==============================================================================
 //
