+++ Quake/gl_ktx2.c
@@ -1,4 +1,5 @@
 #include "quakedef.h"
+#include "glquake.h"
 #include "gl_ktx2.h"
 #include "gl_texmgr.h"
 #include "basisu_transcoder.h"
