+++ Quake/gl_oit.c
@@ -1,4 +1,5 @@
 #include "quakedef.h"
+#include "glquake.h"
 
 #include "gl_oit.h"
