+++ Quake/gl_refrag.c
@@ -22,6 +22,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 // r_efrag.c
 
 #include "quakedef.h"
+#include "glquake.h"
 
 //===========================================================================
