+++ Quake/gl_rlight.c
@@ -22,6 +22,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 // r_light.c
 
 #include "quakedef.h"
+#include "glquake.h"
 #include "../common/lightgrid.h"
 #include "gl_lightgrid.h"
 #include "r_dlight_pool.h"
