+++ Quake/gl_rmain.c
@@ -22,6 +22,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 // r_main.c
 
 #include "quakedef.h"
+#include "glquake.h"
 #include "cl_postfx.h"
 #include "r_postfx.h"
 #include "r_framegraph.h"
