+++ Quake/gl_rmisc.c
@@ -23,6 +23,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 // r_misc.c
 
 #include "quakedef.h"
+#include "glquake.h"
 #include "gl_lightgrid.h"
 #include "mat_material.h"
 #include "r_dlight_pool.h"
