+++ Quake/gl_shaders.c
@@ -22,6 +22,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 */
 
 #include "quakedef.h"
+#include "glquake.h"
 #include "q_ctype.h"
 #include <assert.h>
 #include <setjmp.h>
