+++ Quake/gl_shadow.c
@@ -1,4 +1,5 @@
 #include "quakedef.h"
+#include "glquake.h"
 
 #include "gl_shadow.h"
