+++ Quake/gl_vidsdl.c
@@ -23,6 +23,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 // gl_vidsdl.c -- SDL GL vid component
 
 #include "quakedef.h"
+#include "glquake.h"
 #include "r_framegraph.h"
 #include "cfgfile.h"
 #include "bgmusic.h"
