+++ Quake/gl_warp.c
@@ -21,5 +21,6 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 //gl_warp.c -- warping animation support
 
 #include "quakedef.h"
+#include "glquake.h"
 
 cvar_t r_waterwarp = {"r_waterwarp", "1", CVAR_ARCHIVE};
