+++ Quake/r_backend.h
@@ -1,82 +1,6 @@
 #ifndef R_BACKEND_H
 #define R_BACKEND_H
 
-#include "q_stdinc.h"
-
-#define R_BACKEND_MAX_PROFILE_SLOTS 128
-
-typedef enum render_blend_factor_e
-{
-	R_BLEND_FACTOR_INVALID = 0,
-	R_BLEND_FACTOR_ZERO,
-	R_BLEND_FACTOR_ONE,
-	R_BLEND_FACTOR_SRC_ALPHA,
-	R_BLEND_FACTOR_ONE_MINUS_SRC_ALPHA,
-	R_BLEND_FACTOR_DST_ALPHA,
-	R_BLEND_FACTOR_ONE_MINUS_DST_ALPHA,
-	R_BLEND_FACTOR_SRC_COLOR,
-	R_BLEND_FACTOR_ONE_MINUS_SRC_COLOR,
-	R_BLEND_FACTOR_DST_COLOR,
-	R_BLEND_FACTOR_ONE_MINUS_DST_COLOR
-} render_blend_factor_t;
-
-typedef enum render_backend_resource_type_e
-{
-	R_BACKEND_RESOURCE_NONE = 0,
-	R_BACKEND_RESOURCE_FRAMEBUFFER,
-	R_BACKEND_RESOURCE_TEXTURE,
-	R_BACKEND_RESOURCE_BUFFER
-} render_backend_resource_type_t;
-
-typedef enum render_backend_resource_slot_e
-{
-	R_BACKEND_RESOURCE_SLOT_NONE = 0,
-	R_BACKEND_RESOURCE_SLOT_SCENE_FBO,
-	R_BACKEND_RESOURCE_SLOT_SCENE_COLOR,
-	R_BACKEND_RESOURCE_SLOT_SCENE_VELOCITY,
-	R_BACKEND_RESOURCE_SLOT_SCENE_DEPTH,
-	R_BACKEND_RESOURCE_SLOT_RESOLVED_SCENE_FBO,
-	R_BACKEND_RESOURCE_SLOT_RESOLVED_SCENE_COLOR,
-	R_BACKEND_RESOURCE_SLOT_RESOLVED_SCENE_VELOCITY,
-	R_BACKEND_RESOURCE_SLOT_COMPOSITE_FBO,
-	R_BACKEND_RESOURCE_SLOT_COMPOSITE_COLOR,
-	R_BACKEND_RESOURCE_SLOT_COMPOSITE_DEPTH,
-	R_BACKEND_RESOURCE_SLOT_SHADOW_SUN_DEPTH,
-	R_BACKEND_RESOURCE_SLOT_VELOCITY,
-	R_BACKEND_RESOURCE_SLOT_COUNT
-} render_backend_resource_slot_t;
-
-typedef struct render_backend_resource_ref_s
-{
-	unsigned char type;
-	unsigned short slot;
-	unsigned short opaque_id;
-} render_backend_resource_ref_t;
-
-typedef enum render_backend_primitive_e
-{
-	R_BACKEND_PRIMITIVE_TRIANGLES = 0,
-	R_BACKEND_PRIMITIVE_TRIANGLE_FAN,
-	R_BACKEND_PRIMITIVE_TRIANGLE_STRIP,
-	R_BACKEND_PRIMITIVE_LINES,
-	R_BACKEND_PRIMITIVE_POINTS
-} render_backend_primitive_t;
-
-typedef enum render_backend_index_type_e
-{
-	R_BACKEND_INDEX_TYPE_UINT16 = 0,
-	R_BACKEND_INDEX_TYPE_UINT32
-} render_backend_index_type_t;
-
-typedef enum render_backend_barrier_bits_e
-{
-	R_BACKEND_BARRIER_NONE = 0,
-	R_BACKEND_BARRIER_TEXTURE_FETCH = 1u << 0,
-	R_BACKEND_BARRIER_SHADER_IMAGE_ACCESS = 1u << 1,
-	R_BACKEND_BARRIER_SHADER_STORAGE = 1u << 2,
-	R_BACKEND_BARRIER_COMMAND = 1u << 3,
-	R_BACKEND_BARRIER_ELEMENT_ARRAY = 1u << 4,
-	R_BACKEND_BARRIER_FRAMEBUFFER = 1u << 5
-} render_backend_barrier_bits_t;
+#include "render_api.h"
 
 #endif
