+++ Quake/r_realtimelight.c
@@ -1,4 +1,5 @@
 #include "quakedef.h"
+#include "glquake.h"
 
 #include "r_realtimelight.h"
 #include "r_dlight_pool.h"
