+++ Quake/r_sanitize.c
@@ -1,4 +1,5 @@
 #include "quakedef.h"
+#include "glquake.h"
 
 #include "r_sanitize.h"
