+++ Quake/r_sprite.c
@@ -22,6 +22,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 //r_sprite.c -- sprite model rendering
 
 #include "quakedef.h"
+#include "glquake.h"
 
 typedef struct spritevert_t {
 	vec3_t		pos;
