+++ Quake/r_tonemap.c
@@ -1,4 +1,5 @@
 #include "quakedef.h"
+#include "glquake.h"
 #include "r_tonemap.h"
 
 #include <math.h>
