+++ Quake/view.c
@@ -22,6 +22,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 // view.c -- player eye positioning
 
 #include "quakedef.h"
+#include "glquake.h"
 #include "cl_iwmusic.h"
 #include "cl_postfx.h"
